hwpoison: check with CLD_KILLED|CLD_DUMPED instead of just CLD_KILLED
tinjpage and ttranshuge can get SIGCHLD(CLD_DUMPED) from their child
processes, but now they only check CLD_KILLED, so tests fail.
This behavior of the kernel might not be wrong, because the defalut
action of the SIGBUS is 'coredump', not 'terminate' (see comments in
include/linux/signal.h).
With this patch, we accept SIGCHLD(CLD_DUMPED) as a correct behavior.

Signed-off-by: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Signed-off-by: Chen Gong <gong.chen@linux.intel.com>